lte: (fixes #2899) Inconsistent type for variable cellForWhichToReportCGI

+++ src/lte/model/lte-rrc-sap.h
@@ -327,7 +327,7 @@ public:
     std::list<uint8_t> blackCellsToRemoveList; ///< black cells to remove list
     std::list<BlackCellsToAddMod> blackCellsToAddModList; ///< black cells to add mod list
     bool haveCellForWhichToReportCGI; ///< have cell for which to report CGI?
-    uint8_t cellForWhichToReportCGI; ///< cell for which to report CGI
+    uint16_t cellForWhichToReportCGI; ///< cell for which to report CGI
   };
 
   /**
